fix(lpse2d): translator keys and defaults as the C++ reads them; LPSE's KAP process; HPE tail and limits
Inventory Phase-1 items A13, A19, A37, A38, A41, A43, A45-A49, A56, A57, and two defaults found
on the way. Each read at its C++ read and use site:

- A37: laser.N.phase is in degrees (LightSolver.cpp:1716-1717); read as radians, 18 shipped decks
  (300.26, 158.63, ...) had wrong relative phases on their multi-beam pumps.
- A38: KAP bandwidth is per beam, laser.N.bandwidth.KAP.frequency (LightSolver.cpp:1743), and its
  phase jumps after random dwell times 2 Exp(1) / dW (computeKapTransitionTime), beams independent
  unless grouped; adept read a global key and jumped every 2 pi / dW. core/kap.py draws LPSE's
  process once per run; the FD row and general injectors now carry it too.
- A13: hpe.VminOverVminPhase (default 0: the whole Maxwellian) sets v_min = r wpe h / (2 pi) in vte
  (ElectronTracker.cu:3082, 3375-3382); adept kept a 2.5 vte tail on every deck. v_min = 0 loads,
  normalizes and steps (new tests, 1-D and 2-D).
- A56 / A57: hpe.gammaLimit.* are ZAK rates defaulting to 1e6 (no limit) -- converted to 1/ps;
  thermalizationProbability defaults to (1, 0, 0) (ElectronTracker.cu:3014, 3038-3043).
- A19: Z is a float (ParameterManager.cpp:487), no longer rounded; units model takes floats.
- A41: the PML denominator is laser.SabcDenom (LightSolver.cpp:938-945); the invented
  laser.evolution.abc.SabcDenom is reported.
- A43: lw.landauDamping.enable defaults to false (ParameterManager.cpp:782-787).
- A45 / A46: lw.noise.maxWavenumber maps to noise_max_wavenumber; noise isSymmetric / isAnalytic
  and densityProfile.temporalSlope are reported as unsupported instead of dropped.
- A47: maxBackgroundDensity passes through (default 1.25, up to 1000) instead of being clipped.
- A48: spectral IAWs accept a uniform velocityProfile, as LPSE (IawSolver.cpp:340-356); a varying
  one is refused as LPSE refuses it; fluid.velocity is reported as not a deck key.
- A49: the full Krall-Trivelpiece IAW rate only on the fd path (the spectral step is nu |k|).
- laser.enable defaults to false, raman.solver to static (ParameterManager.cpp:268, 346); a static
  Raman light with its EPW source term is refused as LPSE does.
The shipped decks all set laser.enable, lw.landauDamping.enable, physical.Z and raman.solver, so
only the phases (A37) and the HPE tail / limits (A13, A56, A57) change existing translations.

+++ adept/_lpse2d/datamodel.py
@@ -557,9 +557,9 @@ class TermsModel(BaseModel):
 
 
 class UnitsModel(BaseModel):
-    atomic_number: int
+    atomic_number: float
     envelope_density: float
-    ionization_state: int
+    ionization_state: float
     laser_intensity: str
     laser_wavelength: str
     reference_electron_temperature: str
